Moved manual processing widgets to dnpWidgets
