+++ dnplab/__init__.py
@@ -9,4 +9,5 @@ from .dnpMath import *
 from .dnpHydration import *
 from . import core
 from .dnpResults import *
+from .dnpWidgets import manualphase, manual_align
 from .version import __version__
